Create Count Steps till n greater than 0
Take an integer input n, and you have to do either of these steps.

If the number is even subtract 1 from n

and if the number is odd subtract 3 from n.

Keep on performing these steps till the time the value of n is greater than 0. In the end print the total number of steps performed.