Create Print steps and update maximum
Take n as input from the user. Then you will be given a list of n positive integers, each time you find a new maximumal value, you have to increment the steps by 1.

Take steps as 0 initially and maximum value as -100 in the starting.

In the end print the number of steps performed.